hadron-desktop/python/requirements-training.txt
# Training Pipeline Dependencies
# For QLoRA fine-tuning and GGUF packaging (llama.cpp)

# Core training
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.15.0
trl>=0.7.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0

# Logging & Monitoring
tensorboard>=2.15.0
wandb>=0.16.0

# Data processing
sentencepiece>=0.1.99
protobuf>=3.20.0

# Configuration
pyyaml>=6.0.0

# Utilities
tqdm>=4.66.0
safetensors>=0.4.0
